requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
requests==2.32.4
urllib3==2.4.0
